requirements.txt
fastapi
keras
keras_cv
numpy>=1.22.0
Pillow
pydantic
tensorflow[and-cuda]
tesserocr
uvicorn
opencv-python>=4.5.4
python-multipart